test(core): split capture ws test into own file with correct teardown
