+++ packages/core/test/acceptance/capture.test.js
@@ -4,12 +4,7 @@ const { test, beforeEach } = require('tap');
 const runner = require('../..').runner.runner;
 const nock = require('nock');
 const uuid = require('uuid');
-const csv = require('csv-parse');
-const fs = require('fs');
-const path = require('path');
 const { SSMS } = require('../../lib/ssms');
-const http = require('http');
-const WebSocket = require('ws');
 
 let xmlCapture = null;
 try {
@@ -649,62 +644,6 @@ test('Capture - RegExp', (t) => {
 
       ee.stop().then(() => {
         t.end();
-      });
-    });
-
-    ee.run();
-  });
-});
-
-test('Capture WS - JSON', (t) => {
-  const server = http.createServer();
-  const wss = new WebSocket.Server({ server: server });
-  const targetServer = server.listen(0);
-
-  wss.on('connection', function (ws) {
-    ws.on('message', function (message) {
-      t.match(message, /hello (ws|bar|foo)/, 'matches incoming message');
-      ws.send(JSON.stringify({ foo: 'bar', baz: 'foo' }));
-    });
-  });
-
-  const script = {
-    config: {
-      target: `ws://127.0.0.1:${targetServer.address().port}`,
-      phases: [{ duration: 2, arrivalRate: 5 }]
-    },
-    scenarios: [
-      {
-        engine: 'ws',
-        flow: [
-          {
-            send: { payload: 'hello ws', capture: { json: '$.foo', as: 'foo' } }
-          },
-          { think: 1 },
-          {
-            send: {
-              payload: 'hello {{ foo }}',
-              capture: { json: '$.baz', as: 'baz' }
-            }
-          },
-          { loop: [{ send: 'hello {{ baz }}' }], count: 5 }
-        ]
-      }
-    ]
-  };
-
-  runner(script).then(function (ee) {
-    ee.on('done', (nr) => {
-      const report = SSMS.legacyReport(nr).report();
-
-      t.equal(
-        Object.keys(report.errors).length,
-        0,
-        'There should be no WS errors'
-      );
-
-      ee.stop().then(() => {
-        targetServer.close(t.end);
       });
     });
